test(rm): port tests from backend to api service
- deletion in progress testss
- triggers tests

interface tests are probably unnecessary and a lot of effort would be
needed to make them work, so they were not ported
