+++ apps/astarte_realm_management_api/test/astarte_realm_management_api/devices/devices_test.exs
@@ -25,8 +25,10 @@ defmodule Astarte.RealmManagement.API.DevicesTest do
   alias Astarte.Core.Generators.Device, as: DeviceGenerator
   alias Astarte.DataAccess.Device.DeletionInProgress
   alias Astarte.DataAccess.Repo
+  alias Astarte.DataAccess.Devices.Device
   alias Astarte.DataAccess.Realms.Realm
-  alias Astarte.Core.Device
+  alias Astarte.Core.Device, as: DeviceCore
+  alias Astarte.RealmManagement.API.DeviceRemoval.Queries
 
   describe "deletion in progress tests" do
     @describetag :devices
@@ -35,7 +37,7 @@ defmodule Astarte.RealmManagement.API.DevicesTest do
       device_id: device_id
     } do
       keyspace = Realm.keyspace_name(realm)
-      {:ok, decoded_id} = Device.decode_device_id(device_id)
+      {:ok, decoded_id} = DeviceCore.decode_device_id(device_id)
 
       assert :ok = Devices.delete_device(realm, device_id)
 
@@ -44,6 +46,61 @@ defmodule Astarte.RealmManagement.API.DevicesTest do
 
       assert decoded_id == deletion.device_id
       refute DeletionInProgress.all_ack?(deletion)
+    end
+
+    property "is not queued for deletion if there are no acks", %{realm: realm} do
+      check all(device_id <- Astarte.Core.Generators.Device.id()) do
+        keyspace = Realm.keyspace_name(realm)
+
+        %Device{
+          device_id: device_id
+        }
+        |> Repo.insert!(prefix: keyspace)
+
+        encoded_device_id = DeviceCore.encode_device_id(device_id)
+        :ok = Devices.delete_device(realm, encoded_device_id)
+
+        assert [] = Queries.retrieve_devices_to_delete!(realm)
+
+        %DeletionInProgress{
+          device_id: device_id,
+          dup_end_ack: false,
+          vmq_ack: false,
+          dup_start_ack: false
+        }
+        |> Repo.delete!(prefix: keyspace)
+      end
+    end
+
+    property "is queued for deletion with all acks", %{realm: realm} do
+      check all(device <- Astarte.Core.Generators.Device.device(interfaces: [])) do
+        keyspace = Realm.keyspace_name(realm)
+
+        %Device{
+          device_id: device.device_id
+        }
+        |> Repo.insert!(prefix: keyspace)
+
+        %DeletionInProgress{
+          device_id: device.device_id,
+          vmq_ack: true,
+          dup_end_ack: true,
+          dup_start_ack: true
+        }
+        |> Repo.insert!(prefix: keyspace)
+
+        assert [deletion] = Queries.retrieve_devices_to_delete!(realm)
+        _ = Repo.delete!(deletion)
+
+        device_id = device.device_id
+
+        assert %DeletionInProgress{
+                 device_id: ^device_id,
+                 vmq_ack: true,
+                 dup_end_ack: true,
+                 dup_start_ack: true
+               } = deletion
+      end
     end
 
     property "does not delete a non existing device", %{realm: realm} do
